Replace the conversation cache with a one-shot create-to-load handoff
