+++ src/routes/+page.svelte
@@ -10,7 +10,7 @@
 	import { ERROR_MESSAGES, error } from "$lib/stores/errors";
 	import { storePendingFiles } from "$lib/utils/pendingFiles";
 	import superjson from "superjson";
-	import { setCachedConversation, type ConversationData } from "$lib/utils/conversationCache";
+	import { seedPendingConversation, type ConversationData } from "$lib/utils/pendingConversation";
 	import { useSettingsStore } from "$lib/stores/settings.js";
 	import { useConversationsStore } from "$lib/stores/conversations.svelte";
 	import { findCurrentModel } from "$lib/utils/models";
@@ -78,12 +78,12 @@
 
 			const { conversationId, conversation } = await res.json();
 
-			// The create response embeds the conversation payload; seed the client
-			// cache with it so the conversation page load skips its GET and the
-			// first generation request starts one round-trip sooner.
+			// The create response embeds the conversation payload; hand it to the
+			// conversation page as a one-shot seed so its load skips the GET and
+			// the first generation request starts one round-trip sooner.
 			if (typeof conversation === "string") {
 				try {
-					setCachedConversation(conversationId, superjson.parse<ConversationData>(conversation));
+					seedPendingConversation(conversationId, superjson.parse<ConversationData>(conversation));
 				} catch {
 					// Malformed seed: the page load falls back to a normal fetch.
 				}
